third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/report-only-report-uri-missing-expected.txt

CONSOLE ERROR: The Content Security Policy 'script-src 'unsafe-inline';' was delivered in report-only mode, but does not specify a 'report-uri'; the policy will have no effect. Please either add a 'report-uri' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE ERROR: [Report Only] Refused to load the script 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/report-test.js' because it violates the following Content Security Policy directive: "script-src 'unsafe-inline'".

This test passes if a console message is present, warning about the missing 'report-uri' directive.
